Callbacks data structure

+++ src/main/java/stoufexis/jarpc/client/ClientCallback.java
@@ -1,7 +1,7 @@
 package stoufexis.jarpc.client;
 
 public interface ClientCallback {
-  void onClientDecodeError(long correlationId, RuntimeException exception);
+  void onClientDecodeError(int correlationId, RuntimeException exception);
 
-  void onServerDecodeError(long correlationId);
+  void onServerDecodeError(int correlationId);
 }
